Authentication Update

+++ src/main/java/com/yourbookapp/bookapp/data/UserRepository.java
@@ -1,9 +1,10 @@
 package com.yourbookapp.bookapp.data;
 
 import com.yourbookapp.bookapp.models.User;
+import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     User findByUsername(String username);
 }
